requirements.txt
Pillow # PIL
opencv-python
torch
torchvision
numpy
tqdm
pandas
matplotlib
flask
apscheduler
requests
mysql-connector-python
pyyaml
scipy
seaborn